+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Import tasks from ics: upload a calendar file in *Administration → Periodic Tasks*, pick a project per row and create the periodic tasks **requires migration** ([#182](https://github.com/jperelli/Redmine-Periodic-Task/issues/182)) (@jperelli)
+
 ## v7.1.1 - 2026-09-11
 
 ### Fixes